+++ fbw-a380x/src/systems/instruments/src/MFD/pages/common/FmsPage.tsx
@@ -1,11 +1,37 @@
 import { FlightPlanIndex } from '@fmgc/flightplanning/FlightPlanManager';
-import { ConsumerSubject, DisplayComponent, FSComponent, Subject, Subscription, VNode } from '@microsoft/msfs-sdk';
+import {
+  ConsumerSubject,
+  DisplayComponent,
+  FSComponent,
+  MappedSubject,
+  Subject,
+  SubscribableMapFunctions,
+  Subscription,
+  VNode,
+} from '@microsoft/msfs-sdk';
 import { FmgcFlightPhase } from '@shared/flightphase';
 import { AbstractMfdPageProps } from 'instruments/src/MFD/MFD';
 import { NXSystemMessages } from 'instruments/src/MFD/shared/NXSystemMessages';
 import { ActivePageTitleBar } from 'instruments/src/MFD/pages/common/ActivePageTitleBar';
 import { MfdSimvars } from 'instruments/src/MFD/shared/MFDSimvarPublisher';
 import { FlightPlanEvents } from '@fmgc/flightplanning/sync/FlightPlanEvents';
+import { MfdSystem } from './MfdUiService';
+import {
+  dataStatusUri,
+  fuelAndLoadPage,
+  flightPlanUriPage,
+  lateralRevisionHoldPage,
+  performancePage,
+  secIndexPageUri,
+  initPage,
+  dirToUri,
+  airwaysPage,
+  departurePage,
+  arrivalPage,
+  lateralRevisionPage,
+  verticalRevisionPage,
+  fixInfoUri,
+} from '../../shared/utils';
 import { ReadonlyFlightPlan } from '@fmgc/flightplanning/plans/ReadonlyFlightPlan';
 import { AlternateFlightPlan } from '@fmgc/flightplanning/plans/AlternateFlightPlan';
 import { FlightPlanPerformanceData } from '@fmgc/flightplanning/plans/performance/FlightPlanPerformanceData';
@@ -30,9 +56,52 @@ export abstract class FmsPage<T extends AbstractMfdPageProps = AbstractMfdPagePr
 
   protected readonly tmpyActive = Subject.create<boolean>(false);
 
+  /** TMPY is only shown in PERF, FUEL & LOAD, INIT, SEC INDEX, WIND & FPLN + REVISION Pages (lat rev, vert rev, airways, hold, departure, arrival) */
+  private readonly shouldShowTemporaryPageUris = this.props.mfd.uiService.activeUri.map(
+    (uri) =>
+      uri.sys === MfdSystem.Fms &&
+      (uri.page === performancePage ||
+        uri.page === fuelAndLoadPage ||
+        uri.page === initPage ||
+        uri.page === flightPlanUriPage ||
+        uri.page === airwaysPage ||
+        uri.page === departurePage ||
+        uri.page === arrivalPage ||
+        uri.page === lateralRevisionPage ||
+        uri.page === lateralRevisionHoldPage ||
+        uri.page === verticalRevisionPage ||
+        uri.uri === fixInfoUri ||
+        uri.uri === secIndexPageUri ||
+        uri.uri === dirToUri),
+  );
+
+  private readonly displayTmpy = MappedSubject.create(
+    SubscribableMapFunctions.and(),
+    this.shouldShowTemporaryPageUris,
+    this.tmpyActive,
+  );
+
   protected readonly secActive = Subject.create<boolean>(false);
 
   protected readonly eoActive = Subject.create<boolean>(false);
+
+  private readonly penaltyActive = Subject.create<boolean>(false);
+
+  /** Penalty is only displayed in DATA STATUS, FUEL & LOAD, F-PLN, HOLD, ALTERNATE & WHAT IF Pages */
+  private readonly penaltyUri = this.props.mfd.uiService.activeUri.map(
+    (uri) =>
+      uri.sys === MfdSystem.Fms &&
+      (uri.uri === dataStatusUri ||
+        uri.page === fuelAndLoadPage ||
+        uri.page === flightPlanUriPage ||
+        uri.page === lateralRevisionHoldPage),
+  );
+
+  private readonly displayPenalty = MappedSubject.create(
+    SubscribableMapFunctions.and(),
+    this.penaltyUri,
+    this.penaltyActive,
+  );
 
   protected readonly activeFlightPhase = ConsumerSubject.create<FmgcFlightPhase>(
     this.sub.on('flightPhase'),
@@ -43,6 +112,8 @@ export abstract class FmsPage<T extends AbstractMfdPageProps = AbstractMfdPagePr
 
   public onAfterRender(node: VNode): void {
     super.onAfterRender(node);
+
+    this.subs.push(this.penaltyUri, this.displayPenalty, this.shouldShowTemporaryPageUris, this.displayTmpy);
 
     // this.mfdInViewConsumer = sub.on(this.props.mfd.uiService.captOrFo === 'CAPT' ? 'leftMfdInView' : 'rightMfdInView');
 
@@ -95,6 +166,7 @@ export abstract class FmsPage<T extends AbstractMfdPageProps = AbstractMfdPagePr
       this.props.fmcService.masterFmcChanged.sub(() => {
         // Check if master FMC exists, re-route subjects
         this.props.fmcService.master?.fmgc.data.engineOut.pipe(this.eoActive);
+        this.props.fmcService.master?.fmgc.data.fuelPenaltyActive.pipe(this.penaltyActive);
       }, true),
     );
 
@@ -251,7 +323,9 @@ export abstract class FmsPage<T extends AbstractMfdPageProps = AbstractMfdPagePr
         activePage={this.activePageTitle}
         offset={Subject.create('')}
         eoIsActive={this.eoActive}
-        tmpyIsActive={this.tmpyActive}
+        tmpyIsActive={this.displayTmpy}
+        penaltyIsActive={this.displayPenalty}
+        isFmsSubsystemPage={true}
       />
     );
   }