+++ .github/CHANGELOG.md
@@ -61,6 +61,11 @@
 1. [A32NX/FWS] Replace `A/SKID N/WS FAULT` master caution with `A/SKID N/WS OFF` master caution - @FozzieHi (fozzie)
 1. [FMS] Implement MSFS 2024 EFB route sync - @Benjozork (Benjamin Dupont)
 1. [FMS] Use localizer station declination for true courses on approach legs - @BlueberryKing (BlueberryKing)
+1. [A380X/FMS] Add FUEL PENALTY function - @BravoMike99 (bruno_pt99)
+1. [A380X/MFD] Update DATA/STATUS page with newer software layout - @BravoMike99 (bruno_pt99)
+1. [A32NX/FMS] Fix pilot entered managed descent mach being ignored - @BlueberryKing (BlueberryKing)
+1. [A32NX/FWS] Add `PARK BRK ON` master caution - @FozzieHi (fozzie)
+1. [A32NX/GPWS] Implement simulated GPWC in rust systems - @lukecologne (luke)
 1. [A380X/FMS] Add secondary flight plans to A380X - @flogross89 (floridude)
 
 ## 0.14.0